メモリ空間を print できるようにするために、 ltoa の引数を uint64 に修正。
コミットコメントを書きながら、「負の数の対応ができなくなっているじゃん」
と気づいたが、今のところ負の数対応は必要ないので無視することとする。

+++ util/include/string.h
@@ -1,7 +1,8 @@
 #ifndef __STRING_H__
 #define __STRING_H__
+#include <types.h>
 
-unsigned char ltoa(const long l, char* str, const int max_length, const int radix);
-unsigned char ltoa_10(const long l, char* str, const int max_length);
+unsigned char ltoa(const uint64 l, char* str, const int max_length, const uint64 radix);
+unsigned char ltoa_10(const uint64 l, char* str, const int max_length);
 
 #endif // __STRING_H__
